Switch to new wrapper interface
Make node and java compatible with new interface

+++ examples/key_value/wrapped_lib/wrapped_lib.cpp
@@ -79,11 +79,11 @@ public:
 
 static ubinder::CppWrapper<KeyValueStorageLib> CppWrapperInstance;
 
-void OnRequestFunc(uint64_t request, const char* data, size_t dataSize) {
+void OnRequestFunc(uint32_t request, const char* data, size_t dataSize) {
     CppWrapperInstance.onRequest(request, data, dataSize);
 }
 
-void OnResponseFunc(uint64_t request, const char* data, size_t dataSize) {
+void OnResponseFunc(uint32_t request, const char* data, size_t dataSize) {
     CppWrapperInstance.onResponse(request, data, dataSize);
 }
