+++ include/ubinder/function_types.h
@@ -7,7 +7,7 @@ extern "C" {
 
 typedef void (*Notification)(const char* data, size_t dataSize);
 
-typedef void (*RequestResponse)(uint64_t request, const char* data, size_t dataSize);
+typedef void (*RequestResponse)(uint32_t request, const char* data, size_t dataSize);
 
 #ifdef __cplusplus
 }
