+++ src/core/binding.cpp
@@ -10,17 +10,17 @@ namespace ubinder {
         _client = std::make_unique<Endpoint>(
             [this](Message&& msg) {this->_channel._clientToServer.push(std::move(msg)); },
             [this]() {return this->_channel._serverToClient.get(); },
-            [this](uint64_t request, std::vector<uint8_t>&& data) {this->_clientOnRequest(request, (const char*)data.data(), data.size()); },
-            [this](uint64_t request, std::vector<uint8_t>&& data) {this->_clientOnResponse(request, (const char*)data.data(), data.size()); },
+            [this](uint32_t request, std::vector<uint8_t>&& data) {this->_clientOnRequest(request, (const char*)data.data(), data.size()); },
+            [this](uint32_t request, std::vector<uint8_t>&& data) {this->_clientOnResponse(request, (const char*)data.data(), data.size()); },
             [this](std::vector<uint8_t>&& data) {this->_clientOnNotification((const char*)data.data(), data.size()); });
     }
 
-    void Binding::onRequestFromWrapper(uint64_t request, const char* data, size_t dataSize) {
+    void Binding::onRequestFromWrapper(uint32_t request, const char* data, size_t dataSize) {
         std::vector<uint8_t> buffer(data, data + dataSize);
         binding._client->SendRequest(request, std::move(buffer));
     }
 
-    void Binding::onResponseFromWrapper(uint64_t request, const char* data, size_t dataSize) {
+    void Binding::onResponseFromWrapper(uint32_t request, const char* data, size_t dataSize) {
         std::vector<uint8_t> buffer(data, data + dataSize);
         binding._client->SendResponse(request, std::move(buffer));
     }
@@ -31,11 +31,11 @@ namespace ubinder {
     }
     ///////////////////////////////////
 
-    void Binding::SendRequest(uint64_t request, std::vector<uint8_t>&& reqData) {
+    void Binding::SendRequest(uint32_t request, std::vector<uint8_t>&& reqData) {
         _server->SendRequest(request, std::move(reqData));
     }
 
-    void Binding::SendResponse(uint64_t request, std::vector<uint8_t>&& reqData) {
+    void Binding::SendResponse(uint32_t request, std::vector<uint8_t>&& reqData) {
         _server->SendResponse(request, std::move(reqData));
     }
 
@@ -44,8 +44,8 @@ namespace ubinder {
     }
 
     void Binding::Register(
-        std::function<void(uint64_t, std::vector<uint8_t>&&)>&& onRequest,
-        std::function<void(uint64_t, std::vector<uint8_t>&&)>&& onResponse,
+        std::function<void(uint32_t, std::vector<uint8_t>&&)>&& onRequest,
+        std::function<void(uint32_t, std::vector<uint8_t>&&)>&& onResponse,
         std::function<void(std::vector<uint8_t>&&)>&& onNotification) {
         _server = std::make_unique<Endpoint>(
             [this](Message&& msg) {this->_channel._serverToClient.push(std::move(msg)); },
