+++ src/core/binding.h
@@ -20,12 +20,12 @@ namespace ubinder {
 class Binding {
 public:
     Binding(InitWrapperFunction initWrapper);
-    void SendRequest(uint64_t request, std::vector<uint8_t>&& reqData);
-    void SendResponse(uint64_t request, std::vector<uint8_t>&& reqData);
+    void SendRequest(uint32_t request, std::vector<uint8_t>&& reqData);
+    void SendResponse(uint32_t request, std::vector<uint8_t>&& reqData);
     void SendNotification(std::vector<uint8_t>&& notificationData);
     // Next 3 functions will be called by wrapper lib to send some data to binding
-    static void onRequestFromWrapper(uint64_t request, const char* data, size_t dataSize);
-    static void onResponseFromWrapper(uint64_t request, const char* data, size_t dataSize);
+    static void onRequestFromWrapper(uint32_t request, const char* data, size_t dataSize);
+    static void onResponseFromWrapper(uint32_t request, const char* data, size_t dataSize);
     static void onNotificationFromWrapper(const char* data, size_t dataSize);
 
     // Pointer to the function that client is listening to
@@ -35,8 +35,8 @@ public:
     ::Notification _clientOnNotification;
 
     void Register(
-        std::function<void(uint64_t, std::vector<uint8_t>&&)>&& onRequest,
-        std::function<void(uint64_t, std::vector<uint8_t>&&)>&& onResponse,
+        std::function<void(uint32_t, std::vector<uint8_t>&&)>&& onRequest,
+        std::function<void(uint32_t, std::vector<uint8_t>&&)>&& onResponse,
         std::function<void(std::vector<uint8_t>&&)>&& onNotification);
 
     void StartListen();
