+++ src/core/message.h
@@ -13,7 +13,7 @@ enum MESSAGE_TYPE {
 
 struct Message {
     std::vector<uint8_t> data;
-    uint64_t request;
+    uint32_t request;
     MESSAGE_TYPE type;
 };
